arreglando base de datos
