--- TPfinalGit/app/src/main/java/fi/unju/edu/ar/tpfinal/OpenHelper/CategorySQLite.java
@@ -1,27 +1,0 @@
-package fi.unju.edu.ar.tpfinal.OpenHelper;
-
-import android.content.Context;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
-
-/**
- * Created by HSoft on 6/11/2017.
- */
-
-public class CategorySQLite extends SQLiteOpenHelper {
-
-    public CategorySQLite(Context context, String name, SQLiteDatabase.CursorFactory factory, int version) {
-        super(context, name, factory, version);
-    }
-
-    @Override
-    public void onCreate(SQLiteDatabase db) {
-        db.execSQL(SQLiteUtils.CATEGORIES_TABLE);
-    }
-
-    @Override
-    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS categories");
-        onCreate(db);
-    }
-}
