--- TPfinalGit/app/src/main/java/fi/unju/edu/ar/tpfinal/OpenHelper/ProductSQLite.java
@@ -1,28 +1,0 @@
-package fi.unju.edu.ar.tpfinal.OpenHelper;
-
-import android.content.Context;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
-
-/**
- * Created by HSoft on 6/11/2017.
- */
-
-public class ProductSQLite extends SQLiteOpenHelper {
-
-    public ProductSQLite(Context context, String name, SQLiteDatabase.CursorFactory factory, int version) {
-        super(context, name, factory, version);
-    }
-
-    @Override
-    public void onCreate(SQLiteDatabase db) {
-        db.execSQL(SQLiteUtils.PRODUCTS_TABLE);
-    }
-
-    @Override
-    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS products");
-        onCreate(db);
-    }
-
-}
